Add help command, helptext per command instance
Each command has its own helptext within its class file and the help
command will generate a list of commands and provide the helptext for a
specified command when called with a command argument. I want to use this
method in my Warframe bot too but I'll need to do some major refactoring
given that I have a Command created by a class that doesn't extend Command
and a Command that actually introduces 3 commands via its command regex
and a 3-way if statement.