Fix zws
I had already made the changes from the last commit, and I guess I had pushed them to the server without pushing to github so the changes got reverted when I pulled from the repo at work. I had used the command to post the source of itself earlier and just copied from there, forgetting about the zero-width-spaces that would now be present. Re-pasted it from Modbot's bash that DID have the changes on the repo.